requirements.txt
streamlit
openai
chromadb
pypdf
openai
pysqlite3-binary
tiktoken
